requirements.txt
Cython==0.29.13
easydict==1.9
ipdb==0.12.2
ipython==7.9.0
jsonlines==1.2.0
matplotlib==3.1.1
networkx==2.4
numpy==1.17.3
opencv-python==4.1.1.26
pandas==0.25.2
Pillow==6.2.1
protobuf==3.10.0
pycocotools==2.0.0
PyYAML==5.1.2
regex==2019.8.19
requests==2.22.0
scikit-image==0.16.2
scipy==1.3.1
tensorboard==2.0.0
tensorboardX==1.9
tensorflow==2.0.0
torchvision==0.2.1
tqdm==4.36.1
urllib3==1.25.6
boto3
